fix: Repository 변경
- 복합키 없앤 버전

+++ src/main/java/com/writon/admin/domain/repository/challenge/ChallengeFunctionRepository.java
@@ -3,6 +3,6 @@ package com.writon.admin.domain.repository.challenge;
 import com.writon.admin.domain.entity.challenge.ChallengeFunction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChallengeFunctionRepository extends JpaRepository<ChallengeFunction, Integer> {
+public interface ChallengeFunctionRepository extends JpaRepository<ChallengeFunction, Long> {
 
 }
