+++ src/main/java/com/writon/admin/domain/repository/organization/AdminUserRepository.java
@@ -3,6 +3,6 @@ package com.writon.admin.domain.repository.organization;
 import com.writon.admin.domain.entity.organization.AdminUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminUserRepository extends JpaRepository<AdminUser, Integer> {
+public interface AdminUserRepository extends JpaRepository<AdminUser, Long> {
 
 }
