+++ src/main/java/com/writon/admin/domain/repository/organization/PositionRepository.java
@@ -3,6 +3,6 @@ package com.writon.admin.domain.repository.organization;
 import com.writon.admin.domain.entity.organization.Position;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PositionRepository extends JpaRepository<Position, Integer> {
+public interface PositionRepository extends JpaRepository<Position, Long> {
 
 }
