+++ src/main/java/com/writon/admin/domain/repository/question/KeywordRepository.java
@@ -3,6 +3,6 @@ package com.writon.admin.domain.repository.question;
 import com.writon.admin.domain.entity.question.Keyword;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KeywordRepository extends JpaRepository<Keyword, Integer> {
+public interface KeywordRepository extends JpaRepository<Keyword, Long> {
 
 }
