+++ src/main/java/com/writon/admin/domain/repository/satisfaction/SatisfactionRepository.java
@@ -3,6 +3,6 @@ package com.writon.admin.domain.repository.satisfaction;
 import com.writon.admin.domain.entity.satisfaction.Satisfaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SatisfactionRepository extends JpaRepository<Satisfaction, Integer> {
+public interface SatisfactionRepository extends JpaRepository<Satisfaction, Long> {
 
 }
