+++ src/main/java/com/writon/admin/domain/repository/satisfaction/SatisfactionSubjectiveResultRepository.java
@@ -4,6 +4,6 @@ import com.writon.admin.domain.entity.satisfaction.SatisfactionSubjectiveResult;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SatisfactionSubjectiveResultRepository extends
-    JpaRepository<SatisfactionSubjectiveResult, Integer> {
+    JpaRepository<SatisfactionSubjectiveResult, Long> {
 
 }
